Improved __DreamStateError() to report state name as a string, other misc.
Idk why this wasn't default to begin with. If the name has been registered in the FSM Definition with 'def.nameLookup({})' then it will show the state string., otherwise the number and a prompt for the dev to register the names.

Also got started on a test harness... it failed. But I may see about reviving it.

A few other corrections here and there.